HashMap Collisions: Retrieving When we retrieve hash map values, we also need to be aware that different keys could point to the same array index leading us to retrieve the wrong value.
To avoid this, we’ll search through the linked list at an index until
we find a node with a matching key. If we find the node with the
correct key, we’ll return the value; otherwise, we’ll return null.

in this commit: added Tests and fixed typo in retrieve method